+++ app/src/main/java/com/simma/simmaapp/model/cartResponseModel/Currency.kt
@@ -2,7 +2,7 @@ package com.simma.simmaapp.model.cartResponseModel
 
 data class Currency(
     val code: String,
-    val exchangeRate: Int,
-    val margin: Int,
+    val exchangeRate: Double,
+    val margin: Double,
     val name: Name
 )
