+++ app/src/main/java/com/simma/simmaapp/model/myOrdersModel/Delivery.kt
@@ -3,7 +3,7 @@ package com.simma.simmaapp.model.myOrdersModel
 data class Delivery(
     val address: String,
     val currency: String,
-    val fees: Int,
+    val fees: Double,
     val maximumDate: String,
     val minimumDate: String,
     val recipientName: String
