+++ app/src/main/java/com/simma/simmaapp/model/placeOrder/Currency.kt
@@ -2,7 +2,7 @@ package com.vs.simma.model.placeOrder
 
 data class Currency(
     val code: String,
-    val exchangeRate: Int,
+    val exchangeRate: Double,
     val margin: Int,
     val name: Name
 )
