+++ app/src/main/java/com/simma/simmaapp/model/placeOrder/PaymentDetails.kt
@@ -1,9 +1,9 @@
 package com.vs.simma.model.placeOrder
 
 data class PaymentDetails(
-    val IQDGrandTotal: Int,
-    val IQDShippingCost: Int,
-    val IQDTotal: Int,
-    val IQDUnitShipmentFee: Int,
-    val exchangeRate: Int
+    val IQDGrandTotal: Double,
+    val IQDShippingCost: Double,
+    val IQDTotal: Double,
+    val IQDUnitShipmentFee: Double,
+    val exchangeRate: Double
 )
